normitted a good amount of files including split
